Bug fix reconnect "hopefully everything works as inteded now"
Turns out you need a context with cancel in pingPongFunc since the loop
only runs when the ticker fires. This means there is potentially a (for
now 30 second) window where the function does not detect an error when
the connection is closed. Solution is a select statement as it used to
be before

So to summarize:
  - pingPong relies on context with cancel to return when there is an
    error in listen
  - listen relies on an error in ReadMessage() as a result of the
    connection being closed to return when there is an error in pingPong